+++ src/main/java/org/folio/linked/data/mapper/dto/monograph/instance/InstanceMapperUnit.java
@@ -110,7 +110,7 @@ public class InstanceMapperUnit extends TopResourceMapperUnit {
     coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getCarrier(), CARRIER);
     coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getCopyright(), COPYRIGHT);
     coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getWorkReference(), INSTANTIATES);
-    coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getExtent(), EXTENT);
+    coreMapper.addOutgoingEdges(instance, InstanceRequest.class, instanceDto.getExtentV2(), EXTENT);
     instance.setFolioMetadata(new FolioMetadata(instance).setSource(LINKED_DATA));
     instance.setId(hashService.hash(instance));
     return instance;
@@ -118,6 +118,7 @@ public class InstanceMapperUnit extends TopResourceMapperUnit {
 
   private JsonNode getDoc(InstanceRequest dto) {
     var map = new HashMap<String, List<String>>();
+    putProperty(map, PropertyDictionary.EXTENT, dto.getExtent());
     putProperty(map, DIMENSIONS, dto.getDimensions());
     putProperty(map, EDITION, dto.getEdition());
     putProperty(map, PROJECTED_PROVISION_DATE, dto.getProjectProvisionDate());