+++ NEWS.md
@@ -1,5 +1,5 @@
 ## 2.0.0 (IN PROGRESS)
-- Updated API to support Extent
+- Updated API to support Extent [MODLD-390](https://folio-org.atlassian.net/browse/MODLD-390)
 
 ## 1.0.1 (03-12-2025)
 - Initial release